drm/i915: Add timecounter and cyclecounter to dev_priv and check correlation in debugfs
This patch starts timecounter from i915 load and exposes debugfs
interface to read gpu timestamp and get corresponding system
timestamp. With this we can check the accuracy of correlation of
GPU and system times over different periods. It should be noted
that overhead of reading the system time is not considered in
debugfs i915_timestamp_info.

Signed-off-by: Sagar Arun Kamble <sagar.a.kamble@intel.com>